requirements.txt
opensearch-py
python-dotenv
